Added functionality for USB dongle (specify #define USB_DONGLE).
Added error messages when Particle version is not supported for the
chosen ultrasonic platform.

Moved LED state changes to end of master_frame() for PARTICLE_ADDON
and USB_DONGLE.

Ultrasound output on B port is now done nicely.  Does not interfere
with port B pins which are not used for controlling the US output
FETs.