Added firmware version number reporting via USB for dongles.
Note:  relate_stdobject.HEX should always be built with optimisation level 8.
With level 9, the following errors were observed: (1) threshold levels
reported in diagnostic mode were always overwritten with 1 or 0 (2)
"number of transducers returning a valid signal" was always zero when
the transducer at 90 degrees (US2) was the one with the largest signal
strength.