firmware v1.24
PARTICLE_ADDON can now be used with Particle v2.10.

DIY_ADAPTOR devices were not setting their temperature correctly.
This is seemingly because of a compiler bug, so a workaround was added
into TempRead() in DS1621.c.  Also fixed a typo in TempRead() which
might have caused subzero temperatures to be set incorrectly.

The variable i2c_connected in relate_stdobject.c is now a boolean
instead of an unsigned int.

LEDs now indicate US TX and RX correctly for DIY_ADAPTOR and PARTICLE_ADDON.