firmware v1.27
Adjusted BAD_DISCOVERY_AGE and
RTC1HOUR_MAX_SLOTS_TO_WAIT_FOR_TIMESERVER so that dongles will respond
faster if the host interrupts the sync master.  A new master will take
over in about four seconds.

Tidied up the USB functions.  Added a USBInit() function which sets
the tri-state register for the USB pins appropriately.  The code in
USBReceive() is now a little more human-readable.

USB devices now wait a short delay at the beginning of master_frame()
to give the host PC a chance to change its serial port back to
57.6kbps.

Made the naming of the #defines for the case LEDs and the USB pins
more consistent with the rest of the code.