firmware v1.32
Nonce and RF authentication bytes come in from the host least
significant byte first.  Dongles now accept the data in this fashion,
and also inform their hosts of the authentication data they receive
over RF in this fashion also.

Reference measurement taken out of the authentication command that the
host sends to the dongle.  Authentication command is now 35 bytes
long.