firmware v1.33
Dongles do not acknowledge the current round of authentication if they
have not received a valid ultrasonic measurement (number of
transducers > 0) for that round.

The number of rounds and the number of RF bits to send per round is
configurable by the host.  The number of nonce bits to send per round
is #defined as 3.

When one of the dongles is happy that the last authentication round
has completed, then it drops out of authentication mode, and sends a
special "last round acknowledgement" RAU packet the next five times it
transmits trigger packets.  The other, upon receiving this special
packet, drops out of authentication mode, sending no further
acknowledgements.  The last round acknowledgment packet is special in
that it contains only the current round number (most significant bit
is the ack), and no RF authentication data.