firmware v1.35
RelateRFTriggerSlots() changed so that authentication RF data is now
sent out with every trigger packet, and the auth data is added to the
packet, taking priority over host info, measurements, and discovery
information.

The real time clock is only set once, upon power-up.  If the dongle is
interrupted it does not request the time again upon joining the
network.  This has the consequence that RTC1Hour_high_counter may no
longer be correct, but we do not use this value anyway.  (See comments
in r2d_protocol.c.)

Improved the structure of the conditional statements which decide when
to send host info.

The current authentication round is only marked as sent if the trigger
packet was sent successfully.