fixes for re-worked mac interface, and ethbridge
this also adds support for running multiple nodes and coordinates on different
panids and channels.

The Coordinator does an active scan first to get a list of all allocated
channels, it then starts an energy scan on all unallocated channels and starts
on one of them. Note, if there is already a coordinator with the same panid the
same behaviour is shown. This means you will have two network with the same
panid on different channels with no intra-pan communication.

The panid can be configured in contiki-conf.h with SICSLOWPAN_CONF_PANID,
default is 0xbeef.